Removing console logs
