+++ backend/src/index.ts
@@ -1,7 +1,9 @@
 import * as express from "express";
 import { get as clusterSyncGet } from "./clustersync";
 import { cacheRefreshInterval } from "./config";
+import { get as healthGet } from "./health";
 import { fetch as fetchDeployment, update as updateCache } from "./lib/cache";
+import { get as versionGet } from "./version";
 const app = express();
 
 app.use((req, res, next) => {
@@ -22,6 +24,12 @@ app.get("/api/deployments", (req: express.Request, res: express.Response) => {
 });
 app.get("/api/clustersync", (req: express.Request, res: express.Response) => {
   clusterSyncGet(req, res);
+});
+app.get("/api/health", (req: express.Request, res: express.Response) => {
+  healthGet(req, res);
+});
+app.get("/api/version", (req: express.Request, res: express.Response) => {
+  versionGet(req, res);
 });
 
 (async () => {
