+++ backend/src/lib/author.ts
@@ -6,7 +6,7 @@ import { getConfig } from "../config";
 import { IDeploymentData } from "./common";
 
 /**
- * Returns author information
+ * Fetches author information
  *
  * @param deployment Deployment instance
  */
