+++ backend/src/server.ts
@@ -2,7 +2,9 @@ import * as express from "express";
 import * as path from "path";
 import { get as clusterSyncGet } from "./clustersync";
 import { cacheRefreshInterval } from "./config";
+import { get as healthGet } from "./health";
 import { fetch as fetchDeployment, update as updateCache } from "./lib/cache";
+import { get as versionGet } from "./version";
 
 const app = express();
 
@@ -18,6 +20,12 @@ app.get("/api/deployments", (req: express.Request, res: express.Response) => {
 });
 app.get("/api/clustersync", (req: express.Request, res: express.Response) => {
   clusterSyncGet(req, res);
+});
+app.get("/api/health", (req: express.Request, res: express.Response) => {
+  healthGet(req, res);
+});
+app.get("/api/version", (req: express.Request, res: express.Response) => {
+  versionGet(req, res);
 });
 
 // The "catchall" handler: for any request that doesn't
